% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.MgDynamic}
\alias{print.MgDynamic}
\title{Print a Dynamic Measurement Invariance result}
\usage{
\method{print}{MgDynamic}(x, showPlot = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{MgDynamic}.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Print a Dynamic Measurement Invariance result
}
